pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = "This project make to understand about the apache airflow"
authors = ["suraj507 <surawut.jirasaktavee@gmail.com>"]
readme = "README.md"
keywords = ["airflow", "poetry"]


[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
apache-airflow = "2.2.2"
pytest = "^7.1.3"
ccxt = "^1.93.56"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
